Contract test with Pact pass when Envelope budget set type to Integer, as specified in acceptance criteria.

+++ src/main/java/com/ognjen/budgetok/application/Envelope.java
@@ -16,5 +16,5 @@ public class Envelope {
     @Id
     private Long id;
     private String name;
-    private double budget;
+    private int budget;
 }
